requirements.txt
pandas>=2.0.0
yfinance>=0.2.40
pyswisseph>=2.10.3.2
